Fix not correct child arguments, add cd
In previous version there was problem due to close read end of pipe - if that occured, ancestor of command had
problem because of it. Now we handle this error such way, that we kill ancestor command if successors had problems
being executed (due to bad name of executable or arguments).
Cd implement as well as changed prompt.